% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annofusion.R
\name{annofusion}
\alias{annofusion}
\title{A function for annotation and visualization ver20171204}
\usage{
annofusion(
  Output = NULL,
  Refgene = NULL,
  Cosmic = NULL,
  Pfam = NULL,
  Uniprot = NULL
)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
Annotation and visualization.
}
\examples{
annovisual()
}
\keyword{annovisual()}
