% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDMergeFun.R
\name{BEDMergeFun}
\alias{BEDMergeFun}
\title{common_functions}
\usage{
BEDMergeFun(DT)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
