% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callfusion.R
\name{callfusion}
\alias{callfusion}
\title{A function for fusion detection v0.1.6}
\usage{
callfusion(
  CaseBam = NULL,
  ControlBam = NULL,
  TestID = NULL,
  OutputPath = NULL,
  Thread = 1,
  ControlPanel = NULL,
  TargetBed = NULL,
  Refgene = NULL,
  Gap = NULL,
  Reference = NULL,
  AnalysisType = "DP",
  AnalysisUnit = 1000,
  MinMappingQuality = 0,
  SplitCutoff = 2,
  DiscordantCutoff = 3,
  NucDiv = TRUE,
  SplitRatio = 0.7,
  MatchBase = 10,
  Log = FALSE
)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
Fusion detection.
}
\examples{
callfusion()
}
\keyword{callfusion()}
