% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlpanel.R
\name{controlpanel}
\alias{controlpanel}
\title{common_functions}
\usage{
controlpanel(
  ControlBams = c(),
  ID = NULL,
  OutputPath = NULL,
  Thread = 1,
  SplitCutoff = 4
)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
