% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupfun.R
\name{dupfun}
\alias{dupfun}
\title{common_functions}
\usage{
dupfun(no, bed_tmp, bedin)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
