% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFun.R
\name{IndexFun}
\alias{IndexFun}
\title{common_functions}
\usage{
IndexFun(CaseBam, ControlBam)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
