% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfun.R
\name{logfun}
\alias{logfun}
\title{common_functions}
\usage{
logfun(LogVal, bedLogform, bedform, col, txt)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
