% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapfun.R
\name{mapfun}
\alias{mapfun}
\title{common_functions}
\usage{
mapfun(gn, info, bp, gpmap, pfam, ref)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
