% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbasefun.R
\name{mbasefun}
\alias{mbasefun}
\title{common_functions}
\usage{
mbasefun(cigar, base)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
