% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageFun.R
\name{MessageFun}
\alias{MessageFun}
\title{common_functions}
\usage{
MessageFun(step)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
