% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NucDivFun.R
\name{NucDivFun}
\alias{NucDivFun}
\title{common_functions}
\usage{
NucDivFun(seq, conseq)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
