% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputGenFun.R
\name{OutputGenFun}
\alias{OutputGenFun}
\title{common_functions}
\usage{
OutputGenFun(Output, OutputPath, TestID)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
