% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbedfun.R
\name{pbedfun}
\alias{pbedfun}
\title{common_functions}
\usage{
pbedfun(bed, bed.ran, len, medis)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
