% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranfun.R
\name{ranfun}
\alias{ranfun}
\title{common_functions}
\usage{
ranfun(chr, ou)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
