% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reffun.R
\name{reffun}
\alias{reffun}
\title{common_functions}
\usage{
reffun(gn, pos, ref)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
