% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbasefun.ms.R
\name{scbasefun.ms}
\alias{scbasefun.ms}
\title{common_functions}
\usage{
scbasefun.ms(cigar, base)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
