% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbasefun.sm.R
\name{scbasefun.sm}
\alias{scbasefun.sm}
\title{common_functions}
\usage{
scbasefun.sm(cigar, base)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
