% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectfun.R
\name{selectfun}
\alias{selectfun}
\title{common_functions}
\usage{
selectfun(cpo, bedin)
}
\arguments{
\item{Defaults}{to NULL}
}
\description{
common_functions
}
\examples{
common_functions()
}
\keyword{common_functions()}
